reschedule flow
