+++ public/locale/en.json
@@ -397,6 +397,7 @@
   "appointment_created_success": "Appointment created successfully",
   "appointment_details": "Appointment Details",
   "appointment_not_found": "Appointment not found",
+  "appointment_rescheduled_successfully": "Appointment rescheduled successfully",
   "appointment_type": "Appointment Type",
   "appointments": "Appointments",
   "approve": "Approve",
@@ -1719,6 +1720,8 @@
   "required": "Required",
   "required_quantity": "Required Quantity",
   "reschedule": "Reschedule",
+  "reschedule_appointment": "Reschedule Appointment",
+  "rescheduling": "Rescheduling...",
   "resend_otp": "Resend OTP",
   "reset": "Reset",
   "reset_password": "Reset Password",
